daily_pipeline/requirements.txt
requests
python-dotenv
pandas
psycopg2